Fix conflicts and all tests passed

+++ src/Interpreters/DictionariesMetaStoreFDB.cpp
@@ -143,15 +143,15 @@ std::vector<std::string> DictionariesMetaStoreFDB::split(std::string str, std::s
     int pos;
     std::vector<std::string> result;
     str += pattern; /// Extend string for easy operation
-    int size = str.size();
+    int size = static_cast<int>((str).size());
     for (int i = 0; i < size; i++)
     {
-        pos = str.find(pattern, i);
+        pos = static_cast<int>(str.find(pattern, i));
         if (pos < size)
         {
             std::string s = str.substr(i, pos - i);
             result.push_back(s);
-            i = pos + pattern.size() - 1;
+            i = pos + static_cast<int>(pattern.size()) - 1;
         }
     }
     return result;
@@ -173,7 +173,7 @@ std::vector<std::string> DictionariesMetaStoreFDB::splitXml(std::string xml)
             break;
         }
         loc2 += pattern_end.length();
-        begin = loc2;
+        begin = static_cast<int>(loc2);
         std::string result = "<dictionaries>\n" + xml.substr(loc, loc2 - loc) + "\n</dictionaries>";
         dicts.push_back(result);
     }
